django/{{cookiecutter.project_slug}}/requirements.txt
Django={{ cookiecutter.django_version }}
amqp
aniso8601
arrow
asgiref
async-timeout
atomicwrites
attrs
autobahn
Automat
Babel
backcall
billiard
celery
channels
chardet
constantly
coverage
daphne
decorator
django-adminlte2
django-autocomplete-light
django-cors-headers
django-crispy-forms
django-debug-toolbar
django-extensions
django-filter
django-mathfilters
django-phonenumber-field
django-split-settings
djangorestframework
factory-boy
Faker
graphene
graphene-django
graphene-django-extras
graphq
graphql-relay
gunicorn
hyperlink
idna
incremental
ipython
ipython-genutils
jedi
kombu
Markdown
more-itertools
numpy
pandas
parso
pexpect
phonenumberslite
pickleshare
pkg-resources
pluggy
promise
prompt-toolkit
psycopg2
ptyprocess
py
py-moneyed
Pycco
Pygments
PyHamcrest
pystache
pytest
python-dateutil
pytz
Rx
singledispatch
six
smartypants
sqlparse
tblib
tex
traitlets
Twisted
txaio
uuid
vine
wcwidth
Werkzeug
zope.interface
